backend/pyproject.toml
[tool.poetry]
name = "numanibnmazid"
version = "0.1.0"
description = "Numan Ibn Mazid's portfolio application"
authors = ["NumanIbnMazid <numanibnmazid@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.1"
django-environ = "^0.8.1"
psycopg2-binary = "^2.9.3"
whitenoise = "^5.3.0"
django-debug-toolbar = "^3.2.4"
django-compressor = "^3.1"
uvicorn = "^0.16.0"
django-cors-headers = "^3.11.0"
argon2-cffi = "^21.3.0"
django-safedelete = "^1.1.2"
Pillow = "^9.0.0"
django-crispy-forms = "^1.13.0"
crispy-tailwind = "^0.5.0"
django-allauth = "^0.47.0"
django-cleanup = "^5.2.0"
python-memcached = "^1.59"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
